requirements-advanced.txt
# Advanced MCP Requirements
# Install with: pip install -r requirements-advanced.txt

# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.0.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# OCR for screenshots/images
pytesseract>=0.3.10
Pillow>=10.0.0

# Document processing
PyPDF2>=3.0.0
python-docx>=1.0.0

# Optional: Advanced OCR
# easyocr>=1.7.0

# Optional: More document formats
# openpyxl>=3.1.0  # Excel
# python-pptx>=0.6.0  # PowerPoint
